requirements.txt
sentence-transformers
faiss-cpu
google-generativeai
pypdf
fastapi uvicorn
python-dotenv
